% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfit-class.R
\name{autoplot.sc_fit}
\alias{autoplot.sc_fit}
\title{Autoplot method for \code{sc_fit}}
\usage{
\method{autoplot}{sc_fit}(object, ...)
}
\arguments{
\item{object}{An \code{sc_fit}.}

\item{...}{Passed to \code{plot.sc_fit()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Default ggplot2 entry point; returns the ridgeline plot of
\code{beta(Z)}.
}
