% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfit-class.R
\name{coef.sc_fit}
\alias{coef.sc_fit}
\title{Extract DML point estimates from \code{sc_fit}}
\usage{
\method{coef}{sc_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sc_fit}.}

\item{...}{Unused.}
}
\value{
A named numeric vector of DML point estimates \code{theta_hat}.
}
\description{
Extract DML point estimates from \code{sc_fit}
}
