% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity-class.R
\name{format.sc_quantity_bivariate}
\alias{format.sc_quantity_bivariate}
\title{Format method for \code{sc_quantity_bivariate}}
\usage{
\method{format}{sc_quantity_bivariate}(x, digits = 4L, ...)
}
\arguments{
\item{x}{Bivariate quantity.}

\item{digits}{Significant digits.}

\item{...}{Unused.}
}
\value{
A character string.
}
\description{
Format method for \code{sc_quantity_bivariate}
}
