% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfit-class.R
\name{print.sc_fit}
\alias{print.sc_fit}
\title{Print method for \code{sc_fit}}
\usage{
\method{print}{sc_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sc_fit}.}

\item{...}{Unused, present for method compatibility.}
}
\value{
\code{x}, invisibly.
}
\description{
One-line description: call, K, number of respondents, number of
attribute dummies, number of Z covariates, and seed.
}
