% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfit-class.R
\name{print.sc_fit_summary}
\alias{print.sc_fit_summary}
\title{Print method for \code{sc_fit_summary}}
\usage{
\method{print}{sc_fit_summary}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An \code{sc_fit_summary}.}

\item{digits}{Significant digits.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{sc_fit_summary}
}
