% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-helper.R
\name{print.sc_profile}
\alias{print.sc_profile}
\title{Print method for \code{sc_profile}}
\usage{
\method{print}{sc_profile}(x, ...)
}
\arguments{
\item{x}{An \code{sc_profile}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{sc_profile}
}
