% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity-class.R
\name{print.sc_quantity}
\alias{print.sc_quantity}
\title{Print method for \code{sc_quantity}}
\usage{
\method{print}{sc_quantity}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{sc_quantity}.}

\item{digits}{Number of significant digits (default 4).}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{sc_quantity}
}
