% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity-class.R
\name{print.sc_quantity_bivariate}
\alias{print.sc_quantity_bivariate}
\title{Print method for \code{sc_quantity_bivariate}}
\usage{
\method{print}{sc_quantity_bivariate}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An \code{sc_quantity_bivariate}.}

\item{digits}{Significant digits.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{sc_quantity_bivariate}
}
