% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-diagnostic.R
\name{print.sc_quantity_design_diagnostic}
\alias{print.sc_quantity_design_diagnostic}
\title{Print method for \code{sc_quantity_design_diagnostic}}
\usage{
\method{print}{sc_quantity_design_diagnostic}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An \code{sc_quantity_design_diagnostic} object.}

\item{digits}{Significant digits.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{sc_quantity_design_diagnostic}
}
