% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{sconjoint-reexports}
\alias{sconjoint-reexports}
\alias{autoplot}
\title{Re-exported functions from other packages}
\usage{
autoplot(object, ...)
}
\value{
The re-exported \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} generic. When called on
an \code{sc_fit} object it dispatches to \code{\link[=autoplot.sc_fit]{autoplot.sc_fit()}} and returns
a ggplot object.
}
\description{
The \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} generic is re-exported so that \code{autoplot()}
methods for sconjoint objects (see \code{\link[=autoplot.sc_fit]{autoplot.sc_fit()}}) can be called
without attaching ggplot2.
}
\keyword{internal}
