% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-amce.R
\name{print.sc_quantity_validate_amce}
\alias{print.sc_quantity_validate_amce}
\title{Print method for \code{sc_quantity_validate_amce}}
\usage{
\method{print}{sc_quantity_validate_amce}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An \code{sc_quantity_validate_amce}.}

\item{digits}{Significant digits.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{sc_quantity_validate_amce}
}
