% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sconjoint-package.R
\docType{package}
\name{sconjoint-package}
\alias{sconjoint}
\alias{sconjoint-package}
\title{sconjoint: Structural Deep-Learning Estimation for Conjoint Experiments}
\description{
Implements the structural deep-learning estimator for forced-choice
conjoint experiments of Acharya, Hainmueller, and Xu (2026). The
package embeds a deep neural network in a random utility logit so
that preference parameters vary flexibly with respondent
characteristics, and provides double/debiased machine-learning
inference with cross-fitting clustered at the respondent level.
}
\details{
This is an early development release (milestone M1): the package
skeleton is in place but no statistical methods are implemented yet.
See the GitHub repository for the development roadmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yiqingxu.org/packages/sconjoint/}
  \item Report bugs at \url{https://github.com/xuyiqing/sconjoint/issues}
}

}
\author{
\strong{Maintainer}: Yiqing Xu \email{yiqingxu@stanford.edu}

Authors:
\itemize{
  \item Avidit Acharya
  \item Jens Hainmueller
}

}
\keyword{internal}
