% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{sconjoint-reexports}
\alias{sconjoint-reexports}
\alias{autoplot}
\title{Re-exported functions from other packages}
\usage{
autoplot(object, ...)
}
\description{
Re-exported functions from other packages
}
\keyword{internal}
